FIX: bug fixes in single sweep gridding methods; moving methods from the util to the map module
