+++ pyart/map/__init__.py
@@ -11,5 +11,6 @@ from .grid_mapper import example_roi_func_dist  # noqa
 from .grid_mapper import example_roi_func_dist_beam  # noqa
 from .grid_mapper import grid_from_radars  # noqa
 from .grid_mapper import map_to_grid  # noqa
+from .grid_mapper import grid_ppi_sweeps, grid_rhi_sweeps  # noqa
 
 __all__ = [s for s in dir() if not s.startswith("_")]
